LayoutTests/imported/w3c/web-platform-tests/screen-orientation/orientation-reading-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: NotSupportedError: Screen orientation locking is not supported

Harness Error (FAIL), message = Unhandled rejection: Screen orientation locking is not supported

PASS Test screen.orientation properties
FAIL Test screen.orientation default values. assert_true: expected true got false
PASS Test the orientations and associated angles
PASS Test that screen.orientation properties are not writable
PASS Test that screen.orientation is always the same object
TIMEOUT Test that screen.orientation values change if the orientation changes Test timed out

Harness Error (FAIL), message = Unhandled rejection: Screen orientation locking is not supported

PASS Test screen.orientation properties
FAIL Test screen.orientation default values. assert_true: expected true got false
PASS Test the orientations and associated angles
PASS Test that screen.orientation properties are not writable
PASS Test that screen.orientation is always the same object
TIMEOUT Test that screen.orientation values change if the orientation changes Test timed out
